python/requirements.txt
# Amazon Chronos-2 forecasting
chronos-forecasting>=1.2.0
torch>=2.0.0
numpy>=1.24.0
